refactor: move Lambda EventBridge handlers to feature modules with typed args
- Add From<common::Error> for lambda_runtime::Error to eliminate manual map_err boilerplate
- Create timeline/services/handlers.rs with typed handler functions (Post, SpaceCommon)
- Simplify run.rs event_bridge_handler to thin dispatcher (~120 lines removed)
